objclass, osd: dissect PGLSFilter into the osd_types.h header.
