+++ src/crimson/osd/osd.cc
@@ -51,14 +51,6 @@ namespace {
 using ceph::common::local_conf;
 using ceph::os::FuturizedStore;
 
-PGLSFilter::PGLSFilter() : cct(nullptr)
-{
-}
-
-PGLSFilter::~PGLSFilter()
-{
-}
-
 namespace ceph::osd {
 
 OSD::OSD(int id, uint32_t nonce,
